show alert dialog

+++ app/src/main/java/com/appcraft/testapp/app/ui/fragment/saved/SavedFragment.kt
@@ -11,6 +11,10 @@ import com.appcraft.testapp.app.presentation.saved.SavedView
 import com.appcraft.testapp.app.ui.fragment.fromWeb.TvShowXAdapter
 import kotlinx.android.synthetic.main.fragment_from_web.*
 import moxy.presenter.InjectPresenter
+import pro.appcraft.lib.utils.dialog.AlertDialogAction
+import pro.appcraft.lib.utils.dialog.AlertDialogType
+import pro.appcraft.lib.utils.dialog.showAlertDialog
+import java.util.*
 
 class SavedFragment: BaseFragment(R.layout.fragment_saved), SavedView, TvShowXAdapter.OnItemClickListener {
 
@@ -42,10 +46,30 @@ class SavedFragment: BaseFragment(R.layout.fragment_saved), SavedView, TvShowXAd
     }
 
     override fun onFavoriteClick(tvShowItemMP: TvShowItemMP) {
-        presenter.deleteItem(tvShowItemMP)
+        showBookGroupDeletionDialog(tvShowItemMP)
     }
 
     override fun setListInAdapter(list: List<TvShowItemMP>) {
         tvShowXAdapter.submitList(list)
+    }
+
+    private fun showBookGroupDeletionDialog(tvShowItemMP: TvShowItemMP) {
+        requireContext().showAlertDialog(
+            type = AlertDialogType.ALERT_HORIZONTAL_2_OPTIONS_RIGHT_ACCENT,
+            header = getString(R.string.delete_dialog_header),
+            message = String.format(
+                Locale.getDefault(), getString(R.string.delete_item_dialog_message),
+                tvShowItemMP.name
+            ),
+            actions = listOf(
+                AlertDialogAction(getString(R.string.cancel)) {
+                    it.dismiss()
+                },
+                AlertDialogAction(getString(R.string.delete)) {
+                    it.dismiss()
+                    presenter.deleteItem(tvShowItemMP)
+                }
+            )
+        )
     }
 }
